using Math.round() to round the temperature

+++ Weather_App/src/main/java/weather/model/ForecastData.java
@@ -2,10 +2,10 @@ package weather.model;
 
 public class ForecastData {
     private String description;
-    private double temp;
+    private int temp;
     private String icon;
 
-    public ForecastData(String description, double temp, String icon) {
+    public ForecastData(String description, int temp, String icon) {
         this.description = description;
         this.temp = temp;
         this.icon = icon;
@@ -13,10 +13,12 @@ public class ForecastData {
     public String getDescription() {
         return description;
     }
-    public double getTemp() {
+    public int getTemp() {
         return temp;
     }
     public String getIcon() {
         return icon;
     }
+
+
 }
